+++ src/components/index.js
@@ -179,6 +179,10 @@ export {
 } from './spacer';
 
 export {
+  EuiSteps,
+} from './steps';
+
+export {
   EuiTable,
   EuiTableBody,
   EuiTableHeader,